pyproject.toml
[project]
name = "build-bench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.7",
    "mcp>=1.20.0",
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "tree-sitter>=0.25.2",
    "tree-sitter-cpp>=0.23.4",
    "tree-sitter-go>=0.25.0",
    "tree-sitter-java>=0.23.5",
    "tree-sitter-rust>=0.24.0",
    "tree-sitter-typescript>=0.23.2",
]
